Remove stevedore in favor of serialize by inspect
The issue with stevedore in this use case is that we need to both
serialize and deserialize.  Stevedore only allows us to go one
direction... there is no way to get the entrypoint name from the class
itself.  Doing it this way makes the assumption that the module is
installed (and class) is installed the same way on both sides of the
serialization, which is probably a good assumption to make.