Refactor type annotation decorators to use click
This commit rewrites how item_tasks decorators is implemented to use
click primatives rather than custom logic.  The primary advantage is to
the new implementation is that the commandline interface comes with the
item_task annotations for free.  Also, much of the decorating and
conversion logic is handled by click classes directly which
significantly reduces the number of lines of code.

Signed-off-by: Jonathan Beezley <jonathan.beezley@kitware.com>