Change transform/cleanup signatures to accept *args, **kwargs
Currently transforms are expected to act solely on the state defined
in their __init__ function.  While this provides a clean interface, it
may be too restrictive for some transforms (i.e. if a transform needs
access to the task object that is calling it).  Changing these
signatures now hedges against making breaking changes to the API after
there are a lot of these out in the wild.